+++ R/ggpiestats.R
@@ -30,7 +30,8 @@
 #' @inheritParams ggbetweenstats
 #' @inheritParams statsExpressions::contingency_table
 #' @inheritParams theme_ggstatsplot
-#' @inheritParams gghistostats
+#'
+#' @inheritSection statsExpressions::contingency_table Contingency table analyses
 #'
 #' @seealso \code{\link{grouped_ggpiestats}}, \code{\link{ggbarstats}},
 #'  \code{\link{grouped_ggbarstats}}