+++ NEWS.md
@@ -1,9 +1,11 @@
-# ggstatsplot 0.9.5
+# ggstatsplot 0.9.5.9000 (development version)
 
 N.B. All statistical analysis in `{ggstatsplot}` is carried out in
 `{statsExpressions}`. Thus, to see changes related to statistical expressions,
 read the `NEWS` for that package:
 <https://indrajeetpatil.github.io/statsExpressions/news/index.html>
+
+# ggstatsplot 0.9.5
 
 MAJOR CHANGES